Revise some building stuff

+++ scripts/ethereum-devnet/run.sh
@@ -6,8 +6,13 @@ HERE=$(dirname "$0")
 cd "$HERE/../../" # Change to toplevel directory of git repository
 TOPDIR=$(pwd) # Top directory for the git repository
 
-# First, kill any previously existing geth
-killall geth > /dev/null 2>&1 || true
+FORCE=N
+
+function check_alive() {
+    curl -sSf -X POST \
+         -H "Content-Type: application/json" \
+         --data '{"jsonrpc":"2.0", "method": "web3_clientVersion", "params":[], "id":67}' http://localhost:$RPCPORT
+}
 
 GETH_RUNDIR=$TOPDIR/.ethereum
 mkdir -p $GETH_RUNDIR
@@ -17,6 +22,15 @@ PORT=30303
 RPCPORT=8545
 DATADIR=geth-data
 
+if [ "$FORCE" = "Y" ] ; then
+    # First, kill any previously existing geth
+    killall geth > /dev/null 2>&1 || true
+fi
+
+if check_alive ; then
+    exit 0
+fi
+    
 # clean out any existing data dir, thereby resetting eth blockchain state.
 if [ -d $DATADIR ]; then
     rm -rf $DATADIR
@@ -45,9 +59,7 @@ geth --allow-insecure-unlock \
      --ipcpath .ethereum/geth.ipc \
      > $LOGDIR/testnet.log 2>&1 &
 
-while ! curl -sSf -X POST \
-        -H "Content-Type: application/json" \
-        --data '{"jsonrpc":"2.0", "method": "web3_clientVersion", "params":[], "id":67}' http://localhost:$RPCPORT ; do
+while ! check_alive ; do
     echo "Geth not started yet, waiting..."
     sleep 1
 done
